requirements.txt
pandas>=2.0
numpy>=1.23
scikit-learn>=1.3,<1.5
imbalanced-learn>=0.12.2
joblib>=1.3
fastapi>=0.110
uvicorn>=0.23
pydantic>=2.0
pyyaml>=6.0
lightgbm>=4.0.0
shap>=0.44
matplotlib>=3.7
pytest>=7.4
httpx>=0.25
ruff>=0.4
black>=24.3
